refactor: move server module under library and run server from cli
