+++ cli/build.gradle.kts
@@ -8,6 +8,9 @@ application {
 }
 
 dependencies {
+  implementation(projects.library.server)
+  implementation(projects.library.core)
+  implementation(projects.library.sqlite)
   implementation(projects.library.ktor)
   implementation(libs.kotlinx.coroutines.core)
   implementation(libs.ktor.client.cio)
